refactor(gs524n): 烟感电量报警值类型调整

+++ gs524n.go
@@ -19,7 +19,7 @@ type GS524N struct {
 	SMSLimit    int                `bson:"SMSLimit"`
 	Metrics     struct {
 		Battery struct {
-			Alert float64 `bson:"Alert"`
+			Alert int `bson:"Alert"`
 		} `bson:"Battery"`
 	} `bson:"Metrics"`
 }
